Reverted the SparkBaseShim class and introduced a DecimalArithmeticOverrides class
