+++ sql-plugin/src/main/31xdb/scala/com/nvidia/spark/rapids/shims/Spark31XdbShims.scala
@@ -113,7 +113,7 @@ abstract class Spark31XdbShims extends Spark31XdbShimsBase with Logging {
         parent = p, rule = r, doFloatToIntCheck = true, stringToAnsiDate = true))
   }
 
-  def get31xdbExprs: Map[Class[_ <: Expression], ExprRule[_ <: Expression]] = Seq(
+  def getExprs: Map[Class[_ <: Expression], ExprRule[_ <: Expression]] = Seq(
     GpuOverrides.expr[Cast](
       "Convert a column of one type of data into another type",
       new CastChecks(),
@@ -152,10 +152,6 @@ abstract class Spark31XdbShims extends Spark31XdbShimsBase with Logging {
         override def convertToGpu(child: Expression): GpuExpression = GpuAbs(child, false)
       })
   ).map(r => (r.getClassFor.asSubclass(classOf[Expression]), r)).toMap
-
-  override def getExprs: Map[Class[_ <: Expression], ExprRule[_ <: Expression]] = {
-    get31xdbExprs ++ super.getExprs
-  }
 
   override def getExecs: Map[Class[_ <: SparkPlan], ExecRule[_ <: SparkPlan]] = {
     Seq(
