+++ sql-plugin/src/main/320+/scala/com/nvidia/spark/rapids/shims/Spark320PlusShims.scala
@@ -53,7 +53,7 @@ import org.apache.spark.unsafe.types.CalendarInterval
 /**
  * Shim base class that can be compiled with every supported 3.2.0+
  */
-trait Spark320PlusShims extends SparkBaseShim with RebaseShims with Logging {
+trait Spark320PlusShims extends SparkShims with RebaseShims with Logging {
 
   def getWindowExpressions(winPy: WindowInPandasExec): Seq[NamedExpression]
 
@@ -122,7 +122,7 @@ trait Spark320PlusShims extends SparkBaseShim with RebaseShims with Logging {
 
   override def isWindowFunctionExec(plan: SparkPlan): Boolean = plan.isInstanceOf[WindowExecBase]
 
-  def get320Exprs: Map[Class[_ <: Expression], ExprRule[_ <: Expression]] = Seq(
+  override def getExprs: Map[Class[_ <: Expression], ExprRule[_ <: Expression]] = Seq(
     GpuOverrides.expr[Cast](
       "Convert a column of one type of data into another type",
       new CastChecks(),
@@ -233,10 +233,6 @@ trait Spark320PlusShims extends SparkBaseShim with RebaseShims with Logging {
               TypeSig.DAYTIME, TypeSig.numericAndInterval))),
       (windowExpression, conf, p, r) => new GpuWindowExpressionMeta(windowExpression, conf, p, r))
   ).map(r => (r.getClassFor.asSubclass(classOf[Expression]), r)).toMap
-
-  override def getExprs: Map[Class[_ <: Expression], ExprRule[_ <: Expression]] = {
-    get320Exprs ++ super.getExprs
-  }
 
   override def getExecs: Map[Class[_ <: SparkPlan], ExecRule[_ <: SparkPlan]] = {
     Seq(
