+++ sql-plugin/src/main/320until340-all/scala/com/nvidia/spark/rapids/shims/Spark320until340Shims.scala
@@ -19,7 +19,7 @@ import com.nvidia.spark.rapids._
 
 import org.apache.spark.sql.catalyst.expressions.{AnsiCast, Expression}
 
-trait Spark320until340Shims extends SparkBaseShim {
+trait Spark320until340Shims extends SparkShims {
 
   override def ansiCastRule: ExprRule[ _ <: Expression] = {
     GpuOverrides.expr[AnsiCast](
